+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Features and Improvements
 
+- Add the remaining states to the state filter - [#706](https://github.com/PrefectHQ/ui/pull/706)
 - Add Automations page - [#695](https://github.com/PrefectHQ/ui/pull/695)
 - Add Service Accounts and API keys to the UI - [#678](https://github.com/PrefectHQ/ui/pull/678)
 - Add generic markdown messages for action notifications - [#674](https://github.com/PrefectHQ/ui/pull/674)